requirements.txt
mysqlclient
pymongo
psutil
psycopg2-binary
redis
pyyaml
distro